Restructure commands
* Use Thor subcommands and creates new classes for each new subcommand
* Change commands name to be more expressive
* Change tests to reflect the changes
