+++ .rubocop_todo.yml
@@ -1,12 +1,12 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2018-01-12 14:39:45 +0100 using RuboCop version 0.52.1.
+# on 2018-06-28 12:01:20 -0300 using RuboCop version 0.52.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 1
+# Offense count: 2
 # Cop supports --auto-correct.
 # Configuration parameters: Include, TreatCommentsAsGroupSeparators.
 # Include: **/*.gemspec
@@ -35,15 +35,14 @@ Layout/EmptyLineBetweenDefs:
     - 'lib/burndown_chart.rb'
     - 'lib/scrum_board.rb'
 
-# Offense count: 6
+# Offense count: 5
 # Cop supports --auto-correct.
 Layout/EmptyLines:
   Exclude:
     - 'lib/burndown_chart.rb'
     - 'lib/scrum_board.rb'
-    - 'spec/unit/cli_spec.rb'
 
-# Offense count: 7
+# Offense count: 6
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: empty_lines, no_empty_lines
@@ -95,11 +94,11 @@ Layout/ExtraSpacing:
 # SupportedStyles: auto_detection, squiggly, active_support, powerpack, unindent
 Layout/IndentHeredoc:
   Exclude:
-    - 'lib/cli.rb'
     - 'spec/unit/backup_spec.rb'
     - 'spec/unit/burndown_chart_spec.rb'
     - 'spec/unit/card_spec.rb'
-    - 'spec/unit/cli_spec.rb'
+    - 'spec/unit/cli/get_spec.rb'
+    - 'spec/unit/cli/scrum_spec.rb'
     - 'spec/unit/retrieve_data_spec.rb'
     - 'spec/unit/trello_wrapper_spec.rb'
 
@@ -133,17 +132,16 @@ Layout/MultilineMethodCallBraceLayout:
     - 'spec/unit/burndown_chart_spec.rb'
     - 'spec/unit/cli_spec.rb'
 
-# Offense count: 6
+# Offense count: 45
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, IndentationWidth.
 # SupportedStyles: aligned, indented, indented_relative_to_receiver
 Layout/MultilineMethodCallIndentation:
   Exclude:
     - 'lib/burndown_chart.rb'
-    - 'spec/unit/trello_wrapper_spec.rb'
-    - 'spec/unit/board_mock_spec.rb' # Indendation is used to express structure of board
-    - 'spec/unit/burndown_data_spec.rb' # Indendation is used to express structure of board
-    - 'spec/unit/scrum_board_spec.rb' # Indendation is used to express structure of board
+    - 'spec/unit/board_mock_spec.rb'
+    - 'spec/unit/burndown_data_spec.rb'
+    - 'spec/unit/scrum_board_spec.rb'
 
 # Offense count: 2
 # Cop supports --auto-correct.
@@ -154,7 +152,7 @@ Layout/MultilineOperationIndentation:
     - 'lib/backup.rb'
     - 'lib/card.rb'
 
-# Offense count: 18
+# Offense count: 20
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces.
 # SupportedStyles: space, no_space
@@ -180,7 +178,7 @@ Layout/SpaceInsideArrayLiteralBrackets:
     - 'spec/unit/burndown_chart_spec.rb'
     - 'spec/unit/support/webmocks.rb'
 
-# Offense count: 24
+# Offense count: 27
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces, SpaceBeforeBlockParameters.
 # SupportedStyles: space, no_space
@@ -193,7 +191,7 @@ Layout/SpaceInsideBlockBraces:
     - 'spec/unit/scrum_board_spec.rb'
     - 'trollolo.gemspec'
 
-# Offense count: 38
+# Offense count: 4
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces.
 # SupportedStyles: space, no_space, compact
@@ -202,8 +200,6 @@ Layout/SpaceInsideHashLiteralBraces:
   Exclude:
     - 'lib/empty_column.rb'
     - 'spec/unit/burndown_chart_spec.rb'
-    - 'spec/unit/cli_spec.rb'
-    - 'spec/unit/trello_wrapper_spec.rb'
 
 # Offense count: 97
 # Cop supports --auto-correct.
@@ -232,7 +228,7 @@ Layout/TrailingWhitespace:
 # Offense count: 2
 Lint/AmbiguousRegexpLiteral:
   Exclude:
-    - 'spec/unit/cli_spec.rb'
+    - 'spec/unit/cli/scrum_spec.rb'
     - 'spec/unit/scrum/sprint_cleaner_spec.rb'
 
 # Offense count: 1
@@ -258,41 +254,32 @@ Lint/UnusedBlockArgument:
 # Configuration parameters: AllowUnusedKeywordArguments, IgnoreEmptyMethods.
 Lint/UnusedMethodArgument:
   Exclude:
-    - 'lib/cli.rb'
+    - 'lib/cli/burndown.rb'
 
 # Offense count: 1
 Lint/UselessAssignment:
   Exclude:
     - 'spec/unit/support/webmocks.rb'
 
-# Offense count: 6
+# Offense count: 1
 Naming/AccessorMethodName:
   Exclude:
     - 'lib/burndown_data.rb'
-    - 'lib/cli.rb'
 
-# Offense count: 21
+# Offense count: 22
 # Configuration parameters: Blacklist.
 # Blacklist: END, (?-mix:EO[A-Z]{1})
 Naming/HeredocDelimiterNaming:
   Exclude:
-    - 'lib/cli.rb'
+    - 'lib/cli/get.rb'
+    - 'lib/cli/scrum.rb'
     - 'spec/unit/backup_spec.rb'
     - 'spec/unit/burndown_chart_spec.rb'
     - 'spec/unit/card_spec.rb'
-    - 'spec/unit/cli_spec.rb'
+    - 'spec/unit/cli/get_spec.rb'
+    - 'spec/unit/cli/scrum_spec.rb'
     - 'spec/unit/retrieve_data_spec.rb'
     - 'spec/unit/trello_wrapper_spec.rb'
-
-# Offense count: 6
-# Configuration parameters: MinNameLength, AllowNamesEndingInNumbers, AllowedNames, ForbiddenNames.
-Naming/UncommunicativeMethodArgName:
-  Exclude:
-    - 'lib/burndown_chart.rb'
-    - 'lib/cli.rb'
-    - 'lib/scrum/priority_name.rb'
-    - 'lib/scrum/sprint_board.rb'
-    - 'lib/scrum/sprint_planning_board.rb'
 
 # Offense count: 1
 # Cop supports --auto-correct.
@@ -303,7 +290,7 @@ Security/YAMLLoad:
 # Offense count: 1
 Style/ClassVars:
   Exclude:
-    - 'lib/cli.rb'
+    - 'lib/cli_settings.rb'
 
 # Offense count: 3
 # Cop supports --auto-correct.
@@ -311,7 +298,8 @@ Style/ClassVars:
 # SupportedStyles: assign_to_condition, assign_inside_condition
 Style/ConditionalAssignment:
   Exclude:
-    - 'lib/cli.rb'
+    - 'lib/cli/burndown.rb'
+    - 'lib/cli/get.rb'
     - 'spec/unit/support/webmocks.rb'
 
 # Offense count: 13
@@ -321,7 +309,7 @@ Style/DateTime:
     - 'spec/unit/burndown_chart_spec.rb'
     - 'spec/unit/burndown_data_spec.rb'
 
-# Offense count: 22
+# Offense count: 29
 Style/Documentation:
   Enabled: false
 
@@ -337,17 +325,17 @@ Style/GuardClause:
   Exclude:
     - 'lib/burndown_chart.rb'
     - 'lib/card.rb'
-    - 'lib/cli.rb'
+    - 'lib/cli_settings.rb'
     - 'lib/scrum/backlog_mover.rb'
     - 'lib/scrum_board.rb'
     - 'lib/settings.rb'
 
-# Offense count: 6
+# Offense count: 8
 # Cop supports --auto-correct.
 Style/IfUnlessModifier:
   Exclude:
     - 'lib/burndown_chart.rb'
-    - 'lib/cli.rb'
+    - 'lib/cli/burndown.rb'
     - 'lib/scrum/sprint_cleaner.rb'
     - 'spec/integration/integration_spec_helper.rb'
     - 'spec/unit/spec_helper.rb'
@@ -366,7 +354,7 @@ Style/MixinUsage:
 # Offense count: 1
 Style/MultilineBlockChain:
   Exclude:
-    - 'lib/cli.rb'
+    - 'lib/cli/get.rb'
 
 # Offense count: 2
 # Cop supports --auto-correct.
@@ -389,8 +377,3 @@ Style/NumericPredicate:
 Style/PercentLiteralDelimiters:
   Exclude:
     - 'spec/unit/burndown_chart_spec.rb'
-
-# Offense count: 1
-# Cop supports --auto-correct.
-# Configuration parameters: EnforcedStyle, AllowInnerSlashes.
-# SupportedStyles: slashes, percent_r, mixed
