+++ lib/trollolo.rb
@@ -22,6 +22,12 @@ require 'erb'
 
 require_relative 'array'
 require_relative 'version'
+require_relative 'cli_settings'
+require_relative 'cli/set'
+require_relative 'cli/get'
+require_relative 'cli/backup'
+require_relative 'cli/burndown'
+require_relative 'cli/scrum'
 require_relative 'cli'
 require_relative 'settings'
 require_relative 'column'
